refactor(Geotag): consolidate files
