+++ client/src/dialogs/DialogRoot.jsx
@@ -6,7 +6,7 @@ import ErrorBoundary from '../util/ErrorBoundary'
 import AboutDialog from './About'
 import AnalyticsDialog from './Analytics'
 import FeatureFlagDialog from './FeatureFlagDialog'
-import GeotagDialog from './GeotagDialog'
+import GeotagDialog from './Geotag'
 import SaveAsImageDialog from './SaveAsImageDialog'
 import SettingsDialog from './SettingsDialog'
 import SignInDialog from './SignInDialog'
