+++ client/src/dialogs/Geotag/GeotagDialog.scss
